Phase 5: Updated config urls and moved categories fetch part to layout
